# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: MiroFish-Novel API v2 Baseline
  version: 0.1.0
  description: |
    Rebuild contract baseline for operation, error, and health resources.
servers:
  - url: http://127.0.0.1:5102
...
paths:
...
  /api/v2/operations/{operationId}:
    get:
      operationId: getOperation
      tags: [operations]
      summary: Query a durable operation
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/operations/{operationId}/steps/{stepId}:
    get:
      operationId: getOperationStepTrace
      tags: [operations]
      summary: Query a step trace by operation and step id
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - $ref: "#/components/parameters/StepId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Operation with step trace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/operations/{operationId}/events:
    get:
      operationId: listOperationEvents
      tags: [operations]
      summary: Query a persisted operation event log
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Event log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OperationEvent"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/operations/{operationId}/stream:
    get:
      operationId: streamOperationEvents
      tags: [operations]
      summary: Subscribe to operation SSE events
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "501":
          $ref: "#/components/responses/ErrorResponse"
...
  /api/v2/llm/settings:
    get:
      operationId: getLlmSettings
      tags: [llm]
      summary: Query llm facility settings
      parameters:
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: LLM settings snapshot
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/llm/channels:
    post:
      operationId: createLlmChannel
      tags: [llm]
      summary: Create llm channel
      parameters:
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: LLM channel created
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/llm/channels/{channelKey}:
    patch:
      operationId: updateLlmChannel
      tags: [llm]
      summary: Update llm channel
      parameters:
        - $ref: "#/components/parameters/ChannelKey"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: LLM channel updated
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteLlmChannel
      tags: [llm]
      summary: Delete llm channel
      parameters:
        - $ref: "#/components/parameters/ChannelKey"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: LLM channel deleted
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/llm/channels/{channelKey}/model-syncs:
    post:
      operationId: syncLlmModels
      tags: [llm]
      summary: Start llm model sync
      parameters:
        - $ref: "#/components/parameters/ChannelKey"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Model sync accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResource"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/llm/module-bindings/{moduleKey}:
    put:
      operationId: bindLlmModule
      tags: [llm]
      summary: Bind module to llm channel/model
      parameters:
        - $ref: "#/components/parameters/ModuleKey"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Module binding updated
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteLlmModuleBinding
      tags: [llm]
      summary: Delete module binding
      parameters:
        - $ref: "#/components/parameters/ModuleKey"
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: Module binding deleted
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "501":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/llm/activity:
    get:
      operationId: getLlmActivity
      tags: [llm]
      summary: Query llm activity
      parameters:
        - $ref: "#/components/parameters/TraceId"
      responses:
        "200":
          description: LLM activity projection
          content:
            application/json:
              schema:
                type: object
        "501":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
...
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
...
    ChannelKey:
      name: channelKey
      in: path
      required: true
      schema:
        type: string
    ModuleKey:
      name: moduleKey
      in: path
      required: true
      schema:
        type: string
...
    StepId:
      name: stepId
      in: path
      required: true
      schema:
        type: string
...
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Stable error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, details, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
    ErrorResponse:
      type: object
      required: [trace_id, error]
      properties:
        trace_id:
          type: string
        error:
          $ref: "#/components/schemas/ApiError"
    OperationStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    OperationStepStatus:
      type: string
      enum: [pending, running, succeeded, failed, skipped]
    OperationStep:
      type: object
      required: [step_id, stage, status, progress_percent]
      properties:
        step_id:
          type: string
        stage:
          type: string
        status:
          $ref: "#/components/schemas/OperationStepStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        elapsed_ms:
          type: integer
          nullable: true
    OperationEvent:
      type: object
      required: [event_id, operation_id, type, emitted_at, payload]
      properties:
        event_id:
          type: string
        operation_id:
          type: string
        type:
          type: string
          enum:
            - operation.started
            - operation.progress
            - operation.step.updated
            - operation.completed
            - operation.failed
        emitted_at:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    OperationResource:
      type: object
      required:
        - operation_id
        - trace_id
        - status
        - progress_percent
        - retryable
        - steps
      properties:
        operation_id:
          type: string
        trace_id:
          type: string
        status:
          $ref: "#/components/schemas/OperationStatus"
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        retryable:
          type: boolean
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/OperationStep"